fix: .env, file by component
