+++ src/js/views/payrollSettings.js
@@ -1,14 +1,11 @@
 import React from "react";
 import Flux from "@4geeksacademy/react-flux-dash";
-import { store, fetchTemporal, update, updateProfileImage, searchMe, remove } from '../actions.js';
-import { TIME_FORMAT, DATETIME_FORMAT, DATE_FORMAT, NOW } from '../components/utils.js';
+import { store, fetchTemporal, update, searchMe, remove } from '../actions.js';
+import { DATETIME_FORMAT, NOW } from '../components/utils.js';
 import { validator, ValidationError } from '../utils/validation';
-import Dropzone from 'react-dropzone';
 import DateTime from 'react-datetime';
 import moment from 'moment';
-import { Button } from '../components/index';
 import { DeductionExtendedCard, Theme } from '../components/index';
-import PropTypes from 'prop-types';
 import {Notify} from 'bc-react-notifier';
 export const Employer = (data = {}) => {
 
@@ -45,135 +42,6 @@ export const Employer = (data = {}) => {
         }
     };
 };
-
-export const Deduction = (data = {}) => {
-
-    const _defaults = {
-        id: null,
-        name: '',
-        value: null,
-        description: '',
-        type: 'PERCENTAGE',
-        lock: false,
-        serialize: function () {
-
-            const newDeduction = {
-            };
-
-            return Object.assign(this, newDeduction);
-        }
-    };
-
-    let _deduction = Object.assign(_defaults, data);
-    return {
-        validate: () => {
-            if (validator.isEmpty(_deduction.name)) throw new ValidationError('The deduction name cannot be empty');
-            if (!_deduction.value) throw new ValidationError('Deduction cannot be empty');
-            if (validator.isEmpty(_deduction.description)) throw new ValidationError('The deduction description cannot be empty');
-            return _deduction;
-        },
-        defaults: () => {
-            return _defaults;
-        }
-    };
-};
-
-export class Profile extends Flux.DashView {
-
-    constructor() {
-        super();
-        this.state = {
-            employer: Employer().defaults(),
-        };
-    }
-
-    setEmployer(newEmployer) {
-        const employer = Object.assign(this.state.employer, newEmployer);
-        this.setState({ employer });
-    }
-
-    componentDidMount() {
-        fetchTemporal('employers/me', 'current_employer');
-        this.subscribe(store, 'current_employer', (employer) => {
-            this.setState({ employer });
-        });
-
-    }
-
-    render() {
-        return (<div className="p-1 listcontents company-profile">
-            <h1><span id="company_details">Company Details</span></h1>
-            <form>
-                <div className="row mt-2">
-                    <div className="col-6">
-                        <label>Response Time</label>
-                        <p>You answer applications within <span className="text-success">{this.state.employer.response_time} min.</span></p>
-                    </div>
-                    <div className="col-6">
-                        <label>Rating</label>
-                        <p>Talents rated you with <span className="text-success">{this.state.employer.rating} points.</span></p>
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col-12">
-                        <label>Company Logo</label>
-                        {!this.state.editingImage ?
-                            <div className="company-logo" style={{ backgroundImage: `url(${this.state.employer.picture})` }}>
-                                <Button color="primary" size="small" onClick={() => this.setState({ editingImage: true })} icon="pencil" />
-                            </div>
-                            :
-                            <div>
-                                <Dropzone onDrop={acceptedFiles => this.setState({ uploadCompanyLogo: acceptedFiles[0] })}>
-                                    {({ getRootProps, getInputProps }) => (
-                                        <section className="upload-zone">
-                                            <div {...getRootProps()}>
-                                                <input {...getInputProps()} />
-                                                <p>Drop your company logo here, or click me to open the file browser</p>
-                                            </div>
-                                        </section>
-                                    )}
-                                </Dropzone>
-                                <Button onClick={() => this.setState({ editingImage: false })} color="secondary">Cancel</Button>
-                                <Button onClick={() => updateProfileImage(this.state.uploadCompanyLogo)} color="success">Save</Button>
-                            </div>
-                        }
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col-12">
-                        <label>Company Name</label>
-                        <input type="text" className="form-control" value={this.state.employer.title}
-                            onChange={(e) => this.setEmployer({ title: e.target.value })}
-                        />
-                    </div>
-                </div>
-                <div className="row mt-2">
-                    <div className="col-12">
-                        <label>Website</label>
-                        <input type="text" className="form-control" value={this.state.employer.website}
-                            onChange={(e) => this.setEmployer({ website: e.target.value })}
-                        />
-                    </div>
-                </div>
-                <div className="row mt-2">
-                    <div className="col-12">
-                        <label>Bio</label>
-                        <input type="text" className="form-control" value={this.state.employer.bio}
-                            onChange={(e) => this.setEmployer({ bio: e.target.value })}
-                        />
-                    </div>
-                </div>
-                <div className="mt-4 text-right">
-                    <button
-                        type="button"
-                        className="btn btn-primary"
-                        onClick={() => update({ path: 'employers/me', event_name: 'current_employer' }, Employer(this.state.employer).validate().serialize())}
-                    >Save</button>
-                </div>
-            </form>
-        </div>);
-    }
-}
 
 export class PayrollSettings extends Flux.DashView {
 
@@ -388,161 +256,3 @@ export class PayrollSettings extends Flux.DashView {
         </div>);
     }
 }
-
-/**
- * Create deduction
- */
-export const CreateDeduction = ({ 
-    onSave, 
-    onCancel, 
-    onChange, 
-    catalog, 
-    formData, 
-    bar, 
-    error
- }) => {
-        console.log('CreateDeduction formData: ', formData);
-        console.log('CreateDeduction error: ', error);
-        return ( <form>
-            <div className="row">
-                <div className="col-6">
-                    <label>Name:</label>
-                    <input className="form-control"
-                        value={formData.name}
-                        onChange={(e)=> onChange({ name: e.target.value })}
-                    />
-                </div>
-                <div className="col-6">
-                    <label>Deduction</label>
-                    <input type="number" className="form-control"
-                        value={formData.value}
-                        onChange={(e)=> onChange({value: Number(e.target.value)})}
-                    />
-                </div>
-            </div>
-            <div className="row">
-                <div className="col-12">
-                    <label>Description:</label>
-                    <input className="form-control"
-                        value={formData.description}
-                        onChange={(e)=> onChange({ description: e.target.value })}
-                    />
-                </div>
-            </div>
-            {/* <div className="row">
-                <div className="col-1" style={{ flexDirection: "row", display: "flex", alignItems: "center" }}>
-                    <input type="checkbox"
-                        style={{ marginRight: "6px" }}
-                        checked={formData.lock}
-                        onChange={(e)=> onChange({ lock: e.target.checked})}
-                    />
-                    <label>Lock </label>
-                </div>
-            </div> */}
-            <div className="btn-bar">
-                <button 
-                type="button"
-                className="btn btn-success" 
-                onClick={() => onSave({
-                    name: formData.name,
-                    value: formData.value,
-                    type: 'PERCENTAGE',
-                    description: formData.description
-                })}
-                >
-                    Create
-                </button>
-            </div>
-        </form>);
-};
-
-CreateDeduction.propTypes = {
-    error: PropTypes.string,
-    action: PropTypes.string,
-    bar: PropTypes.object,
-    onSave: PropTypes.func.isRequired,
-    onCancel: PropTypes.func.isRequired,
-    onChange: PropTypes.func.isRequired,
-    formData: PropTypes.object,
-    catalog: PropTypes.object //contains the data needed for the form to load
-};
-/**
- * Edit deduction
- */
-export const UpdateDeduction = ({
-    onSave, 
-    onCancel, 
-    onChange, 
-    catalog, 
-    formData, 
-    bar, 
-    error
- }) => {
-        console.log('CreateDeduction formData: ', formData);
-        console.log('CreateDeduction error: ', error);
-        return ( <form>
-            <div className="row">
-                <div className="col-6">
-                    <label>Name:</label>
-                    <input className="form-control"
-                        value={formData.name}
-                        onChange={(e)=> onChange({ name: e.target.value })}
-                    />
-                </div>
-                <div className="col-6">
-                    <label>Deduction</label>
-                    <input type="number" className="form-control"
-                        value={formData.value}
-                        onChange={(e)=> onChange({ value: Number(e.target.value)})}
-                    />
-                </div>
-            </div>
-            <div className="row">
-                <div className="col-12">
-                    <label>Description:</label>
-                    <input className="form-control"
-                        value={formData.description}
-                        onChange={(e)=> onChange({ description: e.target.value })}
-                    />
-                </div>
-            </div>
-            <div className="row">
-                <div className="col-1" style={{ flexDirection: "row", display: "flex", alignItems: "center" }}>
-                    <input 
-                        type="checkbox"
-                        style={{ marginRight: "6px" }}
-                        checked={formData.lock}
-                        onChange={(e)=> onChange({ lock: e.target.checked})}
-                    />
-                    <label>Lock </label>
-                </div>
-            </div>
-            <div className="btn-bar">
-                <button 
-                type="button"
-                className="btn btn-success" 
-                onClick={() => onSave({
-                    id: formData.id,
-                    name: formData.name,
-                    value: formData.value,
-                    lock: formData.lock,
-                    type: 'PERCENTAGE',
-                    description: formData.description
-                })}
-                >
-                    Save
-                </button>
-            </div>
-        </form>);
-};
-
-UpdateDeduction.propTypes = {
-    error: PropTypes.string,
-    action: PropTypes.string,
-    bar: PropTypes.object,
-    onSave: PropTypes.func.isRequired,
-    onCancel: PropTypes.func.isRequired,
-    onChange: PropTypes.func.isRequired,
-    formData: PropTypes.object,
-    catalog: PropTypes.object //contains the data needed for the form to load
-};
