+++ webpack.prod.js
@@ -15,5 +15,8 @@ module.exports = merge(common, {
             }
         }
     }
-  }
+  },
+  plugins: [
+    new Dotenv({ path: './.env.prod'})
+  ]
 });
